refactor(controller): move fns to deckgen and sendconfig
