+++ internal/ingress/controller/controller.go
@@ -23,14 +23,13 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/blang/semver"
 	"github.com/eapache/channels"
-	"github.com/kong/go-kong/kong"
 	"github.com/kong/kubernetes-ingress-controller/internal/ingress/election"
 	"github.com/kong/kubernetes-ingress-controller/internal/ingress/status"
 	"github.com/kong/kubernetes-ingress-controller/internal/ingress/task"
 	configClientSet "github.com/kong/kubernetes-ingress-controller/pkg/client/configuration/clientset/versioned"
 	"github.com/kong/kubernetes-ingress-controller/pkg/parser"
+	"github.com/kong/kubernetes-ingress-controller/pkg/sendconfig"
 	"github.com/kong/kubernetes-ingress-controller/pkg/store"
 	"github.com/kong/kubernetes-ingress-controller/pkg/util"
 	"github.com/sirupsen/logrus"
@@ -44,26 +43,10 @@ import (
 	knativeClientSet "knative.dev/networking/pkg/client/clientset/versioned"
 )
 
-// Kong Represents a Kong client and connection information
-type Kong struct {
-	URL        string
-	FilterTags []string
-	// Headers are injected into every request to Kong's Admin API
-	// to help with authorization/authentication.
-	Client *kong.Client
-
-	InMemory      bool
-	HasTagSupport bool
-	Enterprise    bool
-
-	Version semver.Version
-
-	Concurrency int
-}
-
 // Configuration contains all the settings required by an Ingress controller
 type Configuration struct {
-	Kong
+	sendconfig.Kong
+
 	KongCustomEntitiesSecret string
 
 	KubeClient       clientset.Interface
