+++ pkg/sendconfig/sendconfig_test.go
@@ -1,4 +1,4 @@
-package controller
+package sendconfig
 
 import (
 	"reflect"
